# lib/api-zod/src/generated/types/pinLength.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type PinLength = (typeof PinLength)[keyof typeof PinLength];

export const PinLength = {
  NUMBER_4: 4,
  NUMBER_6: 6,
} as const;