# lib/api-zod/src/generated/types/pinMutationRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PinLength } from "./pinLength";

export interface PinMutationRequest {
  /** @pattern ^\d{4}$|^\d{6}$ */
  currentPin?: string;
  /** @pattern ^\d{4}$|^\d{6}$ */
  newPin: string;
  length: PinLength;
}