# lib/api-zod/src/generated/types/pinMutationResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PinLength } from "./pinLength";

export interface PinMutationResult {
  updated: boolean;
  enabled: boolean;
  length: PinLength;
}