# lib/api-zod/src/generated/types/pinRemovalRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PinRemovalRequest {
  /** @pattern ^\d{4}$|^\d{6}$ */
  currentPin: string;
}