# lib/api-zod/src/generated/types/pinVerifyRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PinVerifyRequest {
  /** @pattern ^\d{4}$|^\d{6}$ */
  pin: string;
}